Add files via upload
Flush code and option has been cleared this option is porbably not necessary must users have told me that they wanted the full report by deafult

The option has been enabled by default for some time now as opposed to were older versions asked you to y/n to continue for each analysis section 

For the future an option for a condensced analysis may not be necessary and the code I removed was no longer really begin used.

Usage of the mass scan in the help output was not explained it changed since issue #28 was closed

Also the edited the PE sections table swaping the position of entropy and md5 of the pe sections. I thought it was more readable and lined up in the table better.